double being printed as an int %d now cast(int)

+++ src/combat/fight.cc
@@ -1364,7 +1364,7 @@ damage( struct char_data * ch, struct char_data * victim, int dam,
     if ( ( af = affected_by_spell( victim, SPELL_LOCUST_REGENERATION ) ) ) {
         // pc caster
         if ( ch && !IS_NPC( ch ) && GET_IDNUM( ch ) == af->modifier ) {
-            sprintf( buf, "You drain %d mana from $N!", (dam * 0.75) );
+            sprintf( buf, "You drain %d mana from $N!", (int)(dam * 0.75) );
             act( buf, FALSE, ch, 0, victim, TO_CHAR );
             GET_MOVE(ch) = MAX(0, GET_MOVE(ch) - (dam >> 2));
             if (GET_MOVE(ch) > 0)
